refactor(ui): split tui preview logic into module
